Integrate AEMM with NTH integration tests (#209)
* use aemm in e2e tests

* renaming ec2-metadata-test-proxy to webhook-test-proxy

* reorder pod installation in e2e and renamings

* intregrate aemm with eks cluster test

* addressing shellcheck, fmt issues

* update AEMM version and address pr comments
