--- config/helm/ec2-metadata-test-proxy/templates/daemonset.yaml
@@ -1,50 +1,0 @@
-{{- if .Values.ec2MetadataTestProxy.create -}}
-{{- $isWindows := (contains "windows" .Values.targetNodeOs) -}}
-{{- $k8sVersion := printf "%s.%s" .Capabilities.KubeVersion.Major .Capabilities.KubeVersion.Minor | replace "+" "" -}}
-{{- $osSelector := (semverCompare "<1.14" $k8sVersion | ternary "beta.kubernetes.io/os" "kubernetes.io/os") -}}
-apiVersion: apps/v1
-kind: DaemonSet
-metadata:
-  name: {{ .Values.ec2MetadataTestProxy.label }}
-  labels:
-    app: {{ .Values.ec2MetadataTestProxy.label }}
-spec:
-  selector:
-    matchLabels:
-      app: {{ .Values.ec2MetadataTestProxy.label }}
-  template:
-    metadata:
-      labels:
-        app: {{ .Values.ec2MetadataTestProxy.label }}
-    spec:
-      serviceAccountName: {{ template "ec2-metadata-test-proxy.serviceAccountName" . }}
-      nodeSelector:
-        {{ $osSelector }}: {{ $isWindows | ternary "windows" "linux" }}
-      {{- if (not $isWindows) }}
-      hostNetwork: true
-      {{- end }}
-      containers:
-      - name: {{ .Values.ec2MetadataTestProxy.label }}
-        image: {{ .Values.ec2MetadataTestProxy.image.repository }}:{{ .Values.ec2MetadataTestProxy.image.tag }}
-        imagePullPolicy: {{ .Values.ec2MetadataTestProxy.image.pullPolicy }}
-        ports:
-        - containerPort: {{ .Values.ec2MetadataTestProxy.port }}
-          hostPort: {{ .Values.ec2MetadataTestProxy.port }}
-        env:
-        - name: INTERRUPTION_NOTICE_DELAY
-          value: {{ .Values.ec2MetadataTestProxy.interruptionNoticeDelay | quote }}
-        - name: PORT
-          value: {{ .Values.ec2MetadataTestProxy.port | quote }}
-        - name: ENABLE_SCHEDULED_MAINTENANCE_EVENTS
-          value: {{ .Values.ec2MetadataTestProxy.enableScheduledMaintenanceEvents | quote }}
-        - name: SCHEDULED_EVENT_STATUS
-          value: {{ .Values.ec2MetadataTestProxy.scheduledEventStatus | quote }}
-        - name: ENABLE_SPOT_ITN
-          value: {{ .Values.ec2MetadataTestProxy.enableSpotITN | quote }}
-        - name: ENABLE_IMDS_V2
-          value: {{ .Values.ec2MetadataTestProxy.enableIMDSV2 | quote }}
-      {{- if .Values.ec2MetadataTestProxy.tolerations }}
-      tolerations:
-{{ toYaml .Values.ec2MetadataTestProxy.tolerations | indent 8 }}
-      {{- end }}
-{{- end -}}
