--- config/helm/ec2-metadata-test-proxy/templates/service.yaml
@@ -1,10 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: {{ .Values.ec2MetadataTestProxy.label }}
-spec:
-  selector:
-    app: {{ .Values.ec2MetadataTestProxy.label }}
-  ports:
-  - port: {{ .Values.ec2MetadataTestProxy.port }}
-    protocol: TCP
