--- config/helm/ec2-metadata-test-proxy/values.yaml
@@ -1,31 +1,0 @@
-
-nameOverride: ""
-fullnameOverride: ""
-priorityClassName: system-node-critical
-podAnnotations: {}
-rbac:
-    pspEnabled: true
-serviceAccount:
-    name: emtp-sa
-    create: true
-    annotations: {}
-# The ec2-metadata-test-proxy is for testing purposes
-ec2MetadataTestProxy:
-    create: true
-    interruptionNoticeDelay: 15
-    port: 1338
-    label: ec2-metadata-test-proxy
-    scheduledEventStatus: active
-    enableScheduledMaintenanceEvents: false
-    enableSpotITN: true
-    enableIMDSV2: false
-    image:
-        repository: ec2-metadata-test-proxy
-        tag: customtest
-        pullPolicy: IfNotPresent
-    tolerations: []
-regularPodTest:
-    create: true
-    label: regular-pod-test
-    port: 1339
-targetNodeOs: "linux"
