+++ config/helm/webhook-test-proxy/Chart.yaml
@@ -1,12 +1,12 @@
 apiVersion: v1
-name: ec2-metadata-test-proxy
-description: A Helm chart for the AWS EC2 Metadata Test Proxy
+name: webhook-test-proxy
+description: A Helm chart for the webhook test proxy (used in e2e tests)
 version: 0.3.1
 appVersion: 1.0.0
 home: https://github.com/aws/aws-node-termination-handler
 icon: https://raw.githubusercontent.com/aws/eks-charts/master/docs/logo/aws.png
 sources:
-  - https://github.com/aws/aws-node-termination-handler/test/ec2-metadata-test-proxy
+  - https://github.com/aws/aws-node-termination-handler/test/webhook-test-proxy
 maintainers:
   - name: Brandon Wagner
     url: https://github.com/bwagner5
@@ -14,9 +14,6 @@ maintainers:
   - name: Jillian Montalvo
     url: https://github.com/jillmon
     email: jillmon@users.noreply.github.com
-  - name: Matthew Becker
-    url: https://github.com/mattrandallbecker
-    email: mattrandallbecker@users.noreply.github.com
 keywords:
   - eks
   - ec2
