+++ config/helm/webhook-test-proxy/templates/_helpers.tpl
@@ -2,7 +2,7 @@
 {{/*
 Expand the name of the chart.
 */}}
-{{- define "ec2-metadata-test-proxy.name" -}}
+{{- define "webhook-test-proxy.name" -}}
 {{- default .Chart.Name .Values.nameOverride | trunc 63 | trimSuffix "-" -}}
 {{- end -}}
 
@@ -11,7 +11,7 @@ Create a default fully qualified app name.
 We truncate at 63 chars because some Kubernetes name fields are limited to this (by the DNS naming spec).
 If release name contains chart name it will be used as a full name.
 */}}
-{{- define "ec2-metadata-test-proxy.fullname" -}}
+{{- define "webhook-test-proxy.fullname" -}}
 {{- if .Values.fullnameOverride -}}
 {{- .Values.fullnameOverride | trunc 63 | trimSuffix "-" -}}
 {{- else -}}
@@ -27,11 +27,11 @@ If release name contains chart name it will be used as a full name.
 {{/*
 Common labels
 */}}
-{{- define "ec2-metadata-test-proxy.labels" -}}
-app.kubernetes.io/name: {{ include "ec2-metadata-test-proxy.name" . }}
-helm.sh/chart: {{ include "ec2-metadata-test-proxy.chart" . }}
+{{- define "webhook-test-proxy.labels" -}}
+app.kubernetes.io/name: {{ include "webhook-test-proxy.name" . }}
+helm.sh/chart: {{ include "webhook-test-proxy.chart" . }}
 app.kubernetes.io/instance: {{ .Release.Name }}
-k8s-app: ec2-metadata-test-proxy
+k8s-app: webhook-test-proxy
 {{- if .Chart.AppVersion }}
 app.kubernetes.io/version: {{ .Chart.AppVersion | quote }}
 {{- end }}
@@ -41,16 +41,16 @@ app.kubernetes.io/managed-by: {{ .Release.Service }}
 {{/*
 Create chart name and version as used by the chart label.
 */}}
-{{- define "ec2-metadata-test-proxy.chart" -}}
+{{- define "webhook-test-proxy.chart" -}}
 {{- printf "%s-%s" .Chart.Name .Chart.Version | replace "+" "_" | trunc 63 | trimSuffix "-" -}}
 {{- end -}}
 
 {{/*
 Create the name of the service account to use
 */}}
-{{- define "ec2-metadata-test-proxy.serviceAccountName" -}}
+{{- define "webhook-test-proxy.serviceAccountName" -}}
 {{- if .Values.serviceAccount.create -}}
-    {{ default (include "ec2-metadata-test-proxy.fullname" .) .Values.serviceAccount.name }}
+    {{ default (include "webhook-test-proxy.fullname" .) .Values.serviceAccount.name }}
 {{- else -}}
     {{ default "default" .Values.serviceAccount.name }}
 {{- end -}}
