+++ config/helm/webhook-test-proxy/templates/clusterrole.yaml
@@ -1,7 +1,7 @@
 kind: ClusterRole
 apiVersion: rbac.authorization.k8s.io/v1
 metadata:
-  name: {{ include "ec2-metadata-test-proxy.fullname" . }}
+  name: {{ include "webhook-test-proxy.fullname" . }}
 rules:
 - apiGroups:
     - ""
@@ -9,4 +9,3 @@ rules:
     - nodes
   verbs:
     - get
-
