+++ config/helm/webhook-test-proxy/templates/clusterrolebinding.yaml
@@ -1,12 +1,12 @@
 kind: ClusterRoleBinding
 apiVersion: rbac.authorization.k8s.io/v1
 metadata:
-  name: {{ include "ec2-metadata-test-proxy.fullname" . }}
+  name: {{ include "webhook-test-proxy.fullname" . }}
 subjects:
 - kind: ServiceAccount
-  name: {{ template "ec2-metadata-test-proxy.serviceAccountName" . }}
+  name: {{ template "webhook-test-proxy.serviceAccountName" . }}
   namespace: {{ .Release.Namespace }}
 roleRef:
   kind: ClusterRole
-  name: {{ include "ec2-metadata-test-proxy.fullname" . }}
+  name: {{ include "webhook-test-proxy.fullname" . }}
   apiGroup: rbac.authorization.k8s.io
