+++ config/helm/webhook-test-proxy/templates/psp.yaml
@@ -2,9 +2,9 @@
 apiVersion: policy/v1beta1
 kind: PodSecurityPolicy
 metadata:
-  name: {{ template "ec2-metadata-test-proxy.fullname" . }}
+  name: {{ template "webhook-test-proxy.fullname" . }}
   labels:
-{{ include "ec2-metadata-test-proxy.labels" . | indent 4 }}
+{{ include "webhook-test-proxy.labels" . | indent 4 }}
   annotations:
     seccomp.security.alpha.kubernetes.io/allowedProfileNames: '*'
 spec:
@@ -33,28 +33,28 @@ spec:
 kind: ClusterRole
 apiVersion: rbac.authorization.k8s.io/v1
 metadata:
-  name: {{ template "ec2-metadata-test-proxy.fullname" . }}-psp
+  name: {{ template "webhook-test-proxy.fullname" . }}-psp
   labels:
-{{ include "ec2-metadata-test-proxy.labels" . | indent 4 }}
+{{ include "webhook-test-proxy.labels" . | indent 4 }}
 rules:
   - apiGroups: ['policy']
     resources: ['podsecuritypolicies']
     verbs:     ['use']
     resourceNames:
-      - {{ template "ec2-metadata-test-proxy.fullname" . }}
+      - {{ template "webhook-test-proxy.fullname" . }}
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: RoleBinding
 metadata:
-  name: {{ template "ec2-metadata-test-proxy.fullname" . }}-psp
+  name: {{ template "webhook-test-proxy.fullname" . }}-psp
   labels:
-{{ include "ec2-metadata-test-proxy.labels" . | indent 4 }}
+{{ include "webhook-test-proxy.labels" . | indent 4 }}
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: ClusterRole
-  name: {{ template "ec2-metadata-test-proxy.fullname" . }}-psp
+  name: {{ template "webhook-test-proxy.fullname" . }}-psp
 subjects:
   - kind: ServiceAccount
-    name: {{ template "ec2-metadata-test-proxy.serviceAccountName" . }}
+    name: {{ template "webhook-test-proxy.serviceAccountName" . }}
     namespace: {{ .Release.Namespace }}
 {{- end }}
