+++ config/helm/webhook-test-proxy/templates/regular-pod-test.yaml
@@ -19,17 +19,22 @@ spec:
     spec:
       nodeSelector:
         {{ $osSelector }}: {{ $isWindows | ternary "windows" "linux" }}
+        {{- with .Values.nodeSelector }}
+        {{- toYaml . | nindent 8 }}
+        {{- end }}
+        {{- with .Values.linuxNodeSelector }}
+        {{- toYaml . | nindent 8 }}
+        {{- end }}
       {{- if (not $isWindows) }}
       securityContext:
         runAsUser: 1000
         runAsGroup: 3000
         fsGroup: 2000
       {{- end }}
-      serviceAccountName: {{ template "ec2-metadata-test-proxy.serviceAccountName" . }}
       containers:
       - name: {{ .Values.regularPodTest.label }}
-        image: {{ .Values.ec2MetadataTestProxy.image.repository }}:{{ .Values.ec2MetadataTestProxy.image.tag }}
-        imagePullPolicy: {{ .Values.ec2MetadataTestProxy.image.pullPolicy }}
+        image: {{ .Values.webhookTestProxy.image.repository }}:{{ .Values.webhookTestProxy.image.tag }}
+        imagePullPolicy: {{ .Values.webhookTestProxy.image.pullPolicy }}
         env:
         - name: PORT
           value: {{ .Values.regularPodTest.port | quote }}
