+++ config/helm/webhook-test-proxy/templates/serviceaccount.yaml
@@ -1,11 +1,11 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: {{ template "ec2-metadata-test-proxy.serviceAccountName" . }}
+  name: {{ template "webhook-test-proxy.serviceAccountName" . }}
   namespace: {{ .Release.Namespace }}
 {{- with .Values.serviceAccount.annotations }}
   annotations:
 {{ toYaml . | indent 4 }}
 {{- end }}
   labels:
-{{ include "ec2-metadata-test-proxy.labels" . | indent 4 }}
+{{ include "webhook-test-proxy.labels" . | indent 4 }}
