+++ config/helm/webhook-test-proxy/values.yaml
@@ -1,0 +1,26 @@
+
+nameOverride: ""
+fullnameOverride: ""
+priorityClassName: system-node-critical
+podAnnotations: {}
+rbac:
+    pspEnabled: true
+serviceAccount:
+    name: wtp-sa
+    create: true
+    annotations: {}
+# The webhook-test-proxy is for testing purposes
+webhookTestProxy:
+  create: true
+  label: webhook-test-proxy
+  port: 1340
+  image:
+    repository: webhook-test-proxy
+    tag: customtest
+    pullPolicy: IfNotPresent
+  tolerations: []
+regularPodTest:
+    create: true
+    label: regular-pod-test
+    port: 1339
+targetNodeOs: "linux"
