+++ test/e2e/webhook-secret-test
@@ -7,8 +7,8 @@ set -euo pipefail
 #   $KUBECONFIG
 #   $NODE_TERMINATION_HANDLER_DOCKER_REPO
 #   $NODE_TERMINATION_HANDLER_DOCKER_TAG
-#   $EC2_METADATA_DOCKER_REPO
-#   $EC2_METADATA_DOCKER_TAG
+#   $WEBHOOK_DOCKER_REPO
+#   $WEBHOOK_DOCKER_TAG
 
 echo "Starting Webhook URL Secret Test for Node Termination Handler"
 
@@ -22,7 +22,7 @@ helm upgrade --install $CLUSTER_NAME-anth $SCRIPTPATH/../../config/helm/aws-node
   --wait \
   --force \
   --namespace kube-system \
-  --set instanceMetadataURL="http://localhost:$IMDS_PORT" \
+  --set instanceMetadataURL="http://$AEMM_URL:$IMDS_PORT" \
   --set image.repository="$NODE_TERMINATION_HANDLER_DOCKER_REPO" \
   --set image.tag="$NODE_TERMINATION_HANDLER_DOCKER_TAG" \
   --set webhookURLSecretName=webhooksecret \
@@ -30,13 +30,16 @@ helm upgrade --install $CLUSTER_NAME-anth $SCRIPTPATH/../../config/helm/aws-node
   --set enableSpotInterruptionDraining="true" \
   --set enableScheduledEventDraining="true"
 
-helm upgrade --install $CLUSTER_NAME-emtp $SCRIPTPATH/../../config/helm/ec2-metadata-test-proxy/ \
+helm upgrade --install $CLUSTER_NAME-aemm $AEMM_DL_URL \
+  --wait \
+  --namespace default \
+  --set servicePort="$IMDS_PORT"
+
+helm upgrade --install $CLUSTER_NAME-emtp $SCRIPTPATH/../../config/helm/webhook-test-proxy/ \
   --wait \
   --force \
   --namespace default \
-  --set ec2MetadataTestProxy.image.repository="$EC2_METADATA_DOCKER_REPO" \
-  --set ec2MetadataTestProxy.image.tag="$EC2_METADATA_DOCKER_TAG" \
-  --set ec2MetadataTestProxy.port="$IMDS_PORT"
+  --set webhookTestProxy.port="$IMDS_PORT"
 
 TAINT_CHECK_CYCLES=15
 TAINT_CHECK_SLEEP=15
